+++ LocalizationManager/Classes/Constants.swift
@@ -1,0 +1,21 @@
+//
+//  Constants.swift
+//  LocalizationManager
+//
+//  Created by Dominik Hadl on 18/10/2018.
+//  Copyright © 2018 Nodes. All rights reserved.
+//
+
+import Foundation
+
+enum Constants {
+    enum Keys {
+        static let previousAcceptLanguage = "LocalizationManager-PAL"
+        static let lastUpdatedDate = "LocalizationManager-LUD"
+        static let currentBestFitLanguage = "LocalizationManager-BFL"
+        static let availableLanguages = "LocalizationManager-AVL"
+        static let defaultLanguage = "LocalizationManager-DFL"
+        static let languageOverride = "LocalizationManager-LGO"
+        static let persistedLocalizationType = "LocalizationManager-PTT"
+    }
+}