+++ LocalizationManager/Supporting Files/LocalizationManager_iOS.h
@@ -1,0 +1,19 @@
+//
+//  LocalizationManager.h
+//  LocalizationManager
+//
+//  Created by Dominik Hadl on 18/10/2018.
+//  Copyright © 2018 Nodes. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for LocalizationManager.
+FOUNDATION_EXPORT double LocalizationManagerVersionNumber;
+
+//! Project version string for LocalizationManager.
+FOUNDATION_EXPORT const unsigned char LocalizationManagerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <LocalizationManager/PublicHeader.h>
+
+