+++ LocalizationManager/Supporting Files/LocalizationManager_tvOS.h
@@ -1,0 +1,19 @@
+//
+//  LocalizationManager_tvOS.h
+//  LocalizationManager-tvOS
+//
+//  Created by Peter Bødskov on 20/08/2019.
+//  Copyright © 2019 Nodes. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for LocalizationManager_tvOS.
+FOUNDATION_EXPORT double LocalizationManager_tvOSVersionNumber;
+
+//! Project version string for LocalizationManager_tvOS.
+FOUNDATION_EXPORT const unsigned char LocalizationManager_tvOSVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <LocalizationManager_tvOS/PublicHeader.h>
+
+